refactor(session-history): split writer and sampler modules
Extract the remaining session-history responsibilities out of the public
facade so queueing/transactions/readback live in a dedicated writer
module and active-session polling/aggregation live in a dedicated sampler
module.

This keeps the storage layer split that was already landed, reduces
session_history.cpp to lifecycle wiring, updates the build source list
for the new translation units, and closes the remaining review item about
the file carrying too many responsibilities at once.
